7.3.1.2
Bug# 2929399 Debugger menu updated with "Finish" button. This will pass the gdb command "finish" to the debugger. The actual command sent is saved to devcpp.ini and can be hijacked if the user wants to mod this to a custom debugger command.

Bug #2945056 - Corrected menu bar when returning from full screen. I need to test this on a multi-monitor system, but don't have one.

Bug #2946053 - Added PNG extension to GetExtension for bitmaps.